backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
Django==5.0.4
django-browser-reload==1.12.1
django-ninja==1.1.0
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PySocks==1.7.1
pytz==2024.1
selenium==4.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
